Complete PlaceOrder separate component and api / useUpdateValidators unit test
